Synchronize the TSC between all cpus on startup and provide a sysctl, debug.ktr.resynchronize, which will (very expensively) resynchronize the TSCs 10 times a second to reduce drift errors to manageable numbers.
The TSC on all the cpus for the purposes of ktrdump's ought to be
synchronizated within ~200ns of each other, assuming all cpus are running
at the same frequency.

This needs more work.  What we really need to do is calculate the drift
for each cpu and apply it dynamically.