After much hair pulling the problem with dual BGE interfaces not coming up has been tracked down to BGE0 sharing an interrupt with ACPI.  The bug is actually in ACPI.  ACPI installs its interrupt vector too early, before the IO APICs have been configured, so the IO pin is not actually turned on.
Then, later, bge0 installs its interrupt on the same IRQ but the IOAPIC is
still not enabled because the interrupt subsystem thinks it has already
enabled it.

For the moment, just hack a fix into kern_intr.c to always reinitialize
the pin, with a big XXX comment describing the problem.

Reported-by: David Rhodus