Add master.passwd as a name of this manual page.
Make more changes of FreeBSD to DragonFly (somehow were missed
in previous update). Note that the HISTORY section is not changed.

This appears to also revert some of the changes from last update
(in March 2004) such as misformatted "Changes in behaviour ..."

Used "Ap s" for apostrophe "s" as documented in mdoc(5) and
suggested by saw@ on the submit list. (But didn't fix all.)

Specifically mention "FreeBSD" in the reference to versions prior
to 2.0.5.

Use .Fx instead of ".Tn FreeBSD".

Add a missing period.

Change "is" to "was" for past tense for one sentence.

I submitted this via submit@ list on Jan. 1.

Note that I did not remove the entire "Changes in behavior from
older versions" section, but modified some. This could be removed
or improved.

Also TODO: document that "*" asterisk has different behaviour in field 2
for passwd versus master.passwd.