Because multiple opens of /dev/tty only issue one actual open to the underlying tty, unconditionally use FREAD|FWRITE for the VOP_OPEN/VOP_CLOSE calls.  Otherwise the modes may not match up on final close.
Unconditionally reference the underlying tty when obtaining the lock for a
read or write of /dev/tty.  If we do not do this then it is possible for
the session reference to be ripped out from under us while we are obtaining
the lock and for the vnode to be reused for another purpose without our
knowledge.

Reported-by: Kyle Butt <kylebutt@gmail.com>