Do not set O_NONBLOCK on a threaded program's descriptors any more.  Instead, use the new system calls to directly issue non-blocking I/O.  Additionally, force blocking I/O for debug output.
This partly solves the problem of programs such as bmake or gmake fork/exec'd
children which happen to be threaded.  The children would set O_NONBLOCK
on e.g. stdin, stdout, and stderr, resulting in unexpected operation if the
unrelated parent program tries to issue a read or write.

Solves: gmake 'write error' problem