Removed $DragonFly$ lines in the examples.  They are incidental to the examples themselves.
The 'normal' $DragonFly$ tag is still there in the source file, but the other
ones would cause trouble if the output was committed as part of the build
process.  i.e. building a plain text version of the handbook would generate
a page that had several $DragonFly$ tags in the page examples, and then
committing that text version of the output would cause troubles when CVS
would try to expand those tags.