Incorporates FreeBSD's updates to this file, which removed a no-longer-used comment section (in v1.3), and (in v1.4):
"Fix template rules which use a variable reference in the value of the match
attribute.  This is not allowed according to "5.3 Defining Template Rules"
in REC-xslt-19991116, XSL Transformations (XSLT) Version 1.0.
The xsltproc utility prior to version 1.1.15 allowed this erroneously
but in the latest version it has been fixed.

Reported by:	scottl"

The version of xsltproc in pkgsrc is newer, so this is the first step in
getting the docs to build using DragonFly-native pkgsrc tools.