MASSIVE reorganization of the device operations vector.  Change cdevsw to dev_ops.  dev_ops is a syslink-compatible operations vector structure similar to the vop_ops structure used by vnodes.
Remove a huge number of instances where a thread pointer is still being
passed as an argument to various device ops and other related routines.
The device OPEN and IOCTL calls now take a ucred instead of a thread pointer,
and the CLOSE call no longer takes a thread pointer.