Add a remote host capability for both the source and target directory specifications.  A specification in the form user@host:path will initiate a ssh link to the remote machine and attempt to run a cpdup slave.
The protocol is not all that fast but it is a good start.  The cpignore
and md5 features are not currently available for remote sources and the
fsmid feature is not currently available for remote targets.  However, the
incremental backup feature (-H) *IS* available.

The code is currently considered to be highly experimental and does
not yet adjust for mismatched byte orderings (but the protocol is designed
to make this possible in the future).