Fix CERT Vulnerability Note VU#697164:
"A flaw exists in the way that some versions of BIND handle recursive queries.
It is possible for a remote attacker to trigger an INSIST failure by sending
enough recursive queries that the response to the query arrives after all the
clients looking for the response have left the recursion queue.

Note that although BIND versions 9.2.x also contain the underlying flaw that
causes this vulnerability, ISC reports that the vulnerability is not exposed by
these versions."

Nevertheless, import the patch from ISC BIND 9.2.6-P1