Check that namecache references to the mount point are no longer present before unmounting a filesystem.  Forced unmounts ignore the check but will print a warning.
This patch is primarily designed to prevent nullfs partitions from being
unmounted while processes are still present within them.  The normal vnode
check does not work for nullfs mounts since nullfs mounts do not hold any
vnodes of their own.

Note that this will cause a warning to be reported for the root filesystem
when rebooting.