Fix a long-standing bug inherited from FreeBSD.  It is possible for a signal sent to a process group to race a fork().  If the signal is received by the process doing the fork() before it finishes hooking up the new child to the process group the new child will not receive the signal.  The result is that ^C does not always kill all the processes in the process group.
The problem occurs on UP systems if fork1() in the kernel blocks prior to
finishing the process setup.  Such blocking can occur in MALLOC.  The problem
can also occur if an interrupt occurs just as the process issues the fork()
system call, or on SMP systems where the signal is sent from a different cpu.

The solution is to use a lockmgr() lock to interlock the pgrp structure when
a signal is being sent to a process group and to have fork() check for
signal pre-conditions and return ERESTART if such conditions exist to force
processing of the pending signals.

NOTE!  BMAKE, MAKE, GNUMAKE ALSO HAVE UNRELATED SIGNALING BUGS.  These
programs improperly install a SIG_IGN for a few microseconds in order to
test the current signal function for SIGINT and various other signals.
If a ^C is sent while *MAKE is in this window, the *MAKE will ignore the
signal.

Reviewed-by: "Simon 'corecode' Schubert" <corecode@fs.ei.tum.de>