Use a temporary rate set to hold current channel's rate set. The rates in the temporary rate set will have their MSB cleared.
The main reason that we do this is to be compatible with some
AP implementations, that check MSB of supported rate octet in
(Extended) Supported Rates ie of probe requests.  If the checking
fails on AP side, these implementations will not send out probe
responses.

This kind of rate checking is against:
- IEEE Std 802.11, 1999 edition subclause 7.3.2.2
- IEEE Std 802.11g-2003 subclause 7.3.2.14