Make kernel_map, buffer_map, clean_map, exec_map, and pager_map direct structural declarations instead of pointers.  Clean up all related code, in particular kmem_suballoc().
Remove the offset calculation for kernel_object.  kernel_object's page
indices used to be relative to the start of kernel virtual memory in order
to improve the performance of VM page scanning algorithms.  The optimization
is no longer needed now that VM objects use Red-Black trees.  Removal of
the offset simplifies a number of calculations and makes the code more
readable.